Refactored timeline chart to use the new votes endpoint
